move conditional logic to conditional container
