+++ editor/src/components/navigator/navigator-item/navigator-item.tsx
@@ -4,13 +4,8 @@ import { jsx } from '@emotion/react'
 import createCachedSelector from 're-reselect'
 import React from 'react'
 import {
-  ConditionalCase,
-  getConditionalCaseCorrespondingToBranchPath,
   getConditionalClausePath,
   getConditionalFlag,
-  isActiveBranchOfConditional,
-  isDefaultBranchOfConditional,
-  isOverriddenConditional,
   maybeConditionalExpression,
 } from '../../../core/model/conditionals'
 import { MetadataUtils } from '../../../core/model/element-metadata-utils'
@@ -23,7 +18,7 @@ import { ElementPath } from '../../../core/shared/project-file-types'
 import { getValueFromComplexMap } from '../../../utils/map'
 import { unless, when } from '../../../utils/react-conditionals'
 import { useKeepReferenceEqualityIfPossible } from '../../../utils/react-performance'
-import { FlexRow, IcnProps, useColorTheme, UtopiaTheme } from '../../../uuiui'
+import { FlexRow, IcnProps, UtopiaTheme, useColorTheme } from '../../../uuiui'
 import { ThemeObject } from '../../../uuiui/styles/theme/theme-helpers'
 import { isEntryAConditionalSlot } from '../../canvas/canvas-utils'
 import { ChildWithPercentageSize } from '../../common/size-warnings'
@@ -31,10 +26,10 @@ import { EditorAction, EditorDispatch } from '../../editor/action-types'
 import * as EditorActions from '../../editor/actions/action-creators'
 import * as MetaActions from '../../editor/actions/meta-actions'
 import {
+  NavigatorEntry,
   defaultElementWarnings,
   isConditionalClauseNavigatorEntry,
   isRegularNavigatorEntry,
-  NavigatorEntry,
   navigatorEntryToKey,
   varSafeNavigatorEntryToKey,
 } from '../../editor/store/editor-state'
@@ -46,7 +41,6 @@ import { ExpandableIndicator } from './expandable-indicator'
 import { ItemLabel } from './item-label'
 import { LayoutIcon } from './layout-icon'
 import { NavigatorItemActionSheet } from './navigator-item-components'
-import { assertNever } from '../../../core/shared/utils'
 
 export function getItemHeight(navigatorEntry: NavigatorEntry): number {
   if (isConditionalClauseNavigatorEntry(navigatorEntry)) {
@@ -88,9 +82,10 @@ export interface NavigatorItemInnerProps {
   selected: boolean
   parentOutline: ParentOutline
   visibleNavigatorTargets: Array<NavigatorEntry>
+  onMouseDown?: (e: React.MouseEvent<HTMLDivElement>) => void
 }
 
-function getSelectionActions(
+export function getSelectionActions(
   getSelectedViewsInRange: (i: number) => Array<ElementPath>,
   index: number,
   elementPath: ElementPath,
@@ -111,47 +106,6 @@ function getSelectionActions(
   } else {
     return []
   }
-}
-
-type ConditionalOverrideUpdate = ConditionalCase | 'clear-override' | 'no-update'
-
-function getConditionalOverrideActions(
-  targetPath: ElementPath,
-  conditionalOverrideUpdate: ConditionalOverrideUpdate,
-): Array<EditorAction> {
-  switch (conditionalOverrideUpdate) {
-    case 'no-update':
-      return []
-    case 'clear-override':
-      return [EditorActions.setConditionalOverriddenCondition(targetPath, null)]
-    case 'true-case':
-      return [EditorActions.setConditionalOverriddenCondition(targetPath, true)]
-    case 'false-case':
-      return [EditorActions.setConditionalOverriddenCondition(targetPath, false)]
-    default:
-      assertNever(conditionalOverrideUpdate)
-  }
-}
-
-function selectItem(
-  dispatch: EditorDispatch,
-  getSelectedViewsInRange: (i: number) => Array<ElementPath>,
-  navigatorEntry: NavigatorEntry,
-  index: number,
-  selected: boolean,
-  event: React.MouseEvent<HTMLDivElement>,
-  conditionalOverrideUpdate: ConditionalOverrideUpdate,
-) {
-  const elementPath = navigatorEntry.elementPath
-  const selectionActions = isConditionalClauseNavigatorEntry(navigatorEntry)
-    ? []
-    : getSelectionActions(getSelectedViewsInRange, index, elementPath, selected, event)
-
-  const conditionalOverrideActions = isConditionalClauseNavigatorEntry(navigatorEntry)
-    ? getConditionalOverrideActions(elementPath, conditionalOverrideUpdate)
-    : getConditionalOverrideActions(EP.parentPath(elementPath), conditionalOverrideUpdate)
-
-  dispatch([...conditionalOverrideActions, ...selectionActions], 'leftpane')
 }
 
 const highlightItem = (
@@ -398,8 +352,7 @@ export const NavigatorItem: React.FunctionComponent<
     selected,
     collapsed,
     navigatorEntry,
-    getSelectedViewsInRange,
-    index,
+    onMouseDown,
   } = props
 
   const colorTheme = useColorTheme()
@@ -463,45 +416,6 @@ export const NavigatorItem: React.FunctionComponent<
     'NavigatorItem isConditional',
   )
 
-  const conditionalOverrideUpdate = useEditorState(
-    Substores.metadata,
-    (store): ConditionalOverrideUpdate => {
-      const path = navigatorEntry.elementPath
-      const metadata = store.editor.jsxMetadata
-      const elementMetadata = MetadataUtils.findElementByElementPath(
-        store.editor.jsxMetadata,
-        navigatorEntry.elementPath,
-      )
-      if (isConditionalClauseNavigatorEntry(navigatorEntry)) {
-        if (isActiveBranchOfConditional(navigatorEntry.clause, elementMetadata)) {
-          if (isOverriddenConditional(elementMetadata)) {
-            return 'clear-override'
-          } else {
-            return navigatorEntry.clause
-          }
-        } else {
-          return navigatorEntry.clause
-        }
-      } else {
-        const conditionalCase = getConditionalCaseCorrespondingToBranchPath(path, metadata)
-        if (conditionalCase != null) {
-          const parentPath = EP.parentPath(path)
-          const parentMetadata = MetadataUtils.findElementByElementPath(metadata, parentPath)
-          if (isActiveBranchOfConditional(conditionalCase, parentMetadata)) {
-            return 'no-update'
-          } else if (isDefaultBranchOfConditional(conditionalCase, parentMetadata)) {
-            return 'clear-override'
-          } else {
-            return conditionalCase
-          }
-        }
-
-        return 'no-update'
-      }
-    },
-    'NavigatorItem conditionalOverrideUpdate',
-  )
-
   const isInsideComponent =
     EP.isInsideFocusedComponent(navigatorEntry.elementPath) || isFocusedComponent
   const fullyVisible = useStyleFullyVisible(navigatorEntry)
@@ -550,19 +464,7 @@ export const NavigatorItem: React.FunctionComponent<
     },
     [dispatch, navigatorEntry.elementPath],
   )
-  const select = React.useCallback(
-    (event: any) =>
-      selectItem(
-        dispatch,
-        getSelectedViewsInRange,
-        navigatorEntry,
-        index,
-        selected,
-        event,
-        conditionalOverrideUpdate,
-      ),
-    [dispatch, getSelectedViewsInRange, navigatorEntry, index, selected, conditionalOverrideUpdate],
-  )
+
   const highlight = React.useCallback(
     () => highlightItem(dispatch, navigatorEntry.elementPath, selected, isHighlighted),
     [dispatch, navigatorEntry.elementPath, selected, isHighlighted],
@@ -632,7 +534,7 @@ export const NavigatorItem: React.FunctionComponent<
       <FlexRow
         data-testid={NavigatorItemTestId(varSafeNavigatorEntryToKey(navigatorEntry))}
         style={rowStyle}
-        onMouseDown={select}
+        onMouseDown={onMouseDown}
         onMouseMove={highlight}
         onDoubleClick={focusComponent}
       >
